Added in a $key and a $batch_id
The $key is used to store additional information in site options by prepending a unique value to the end,  these can then be deleted on completion.

The $batch_id is the option_id where the que is stored,  you can use this to query the currently active que.  Specifically a certain $que.